# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ZGRZYT API",
        "description": "Dokumentacja API dla systemu Zgłoszeń i Rejestru Zdarzeń Technicznych (ZGRZYT). Jest to webowa aplikacja do zgłaszania, obsługi i monitorowania błędów w systemach informatycznych.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tickets": {
            "get": {
                "tags": [
                    "Zgłoszenia"
                ],
                "summary": "Wyświetla listę zgłoszeń",
                "description": "Zwraca listę zgłoszeń. Użytkownicy z rolą 'user' widzą tylko swoje zgłoszenia. Użytkownicy 'it' i 'admin' widzą wszystkie.",
                "operationId": "listTickets",
                "responses": {
                    "200": {
                        "description": "Lista zgłoszeń.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/TicketFull"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Błąd autoryzacji."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Zgłoszenia"
                ],
                "summary": "Zapisuje nowe zgłoszenie",
                "description": "Tworzy nowe zgłoszenie w imieniu zalogowanego użytkownika.",
                "operationId": "createTicket",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Drukarka nie działa",
                                        "maxLength": 255
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Drukarka w pokoju 101 nie drukuje, świeci się czerwona lampka."
                                    },
                                    "priority": {
                                        "type": "string",
                                        "example": "niski",
                                        "enum": [
                                            "niski",
                                            "średni",
                                            "wysoki"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Zgłoszenie pomyślnie utworzone.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ticket"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Błąd walidacji."
                    },
                    "401": {
                        "description": "Błąd autoryzacji."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tickets/{ticket}": {
            "get": {
                "tags": [
                    "Zgłoszenia"
                ],
                "summary": "Wyświetla określone zgłoszenie",
                "description": "Pobiera szczegółowe informacje o pojedynczym zgłoszeniu, w tym dane zgłaszającego, przypisanego pracownika IT oraz historię wiadomości.",
                "operationId": "showTicket",
                "parameters": [
                    {
                        "name": "ticket",
                        "in": "path",
                        "description": "ID zgłoszenia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Szczegóły zgłoszenia.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TicketFull"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Błąd autoryzacji."
                    },
                    "403": {
                        "description": "Brak uprawnień do wyświetlenia tego zgłoszenia."
                    },
                    "404": {
                        "description": "Nie znaleziono zgłoszenia."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Zgłoszenia"
                ],
                "summary": "Aktualizuje określone zgłoszenie",
                "description": "Aktualizuje status, priorytet lub przypisanie pracownika IT do zgłoszenia. Dostępne tylko dla ról 'it' i 'admin'.",
                "operationId": "updateTicket",
                "parameters": [
                    {
                        "name": "ticket",
                        "in": "path",
                        "description": "ID zgłoszenia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pola do zaktualizowania.",
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "nowe",
                                            "w trakcie",
                                            "zamknięte"
                                        ]
                                    },
                                    "priority": {
                                        "type": "string",
                                        "enum": [
                                            "niski",
                                            "średni",
                                            "wysoki"
                                        ]
                                    },
                                    "assigned_it_id": {
                                        "description": "ID użytkownika (z rolą IT/Admin) do przypisania.",
                                        "type": "integer",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Zgłoszenie pomyślnie zaktualizowane.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Ticket"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Błąd autoryzacji."
                    },
                    "403": {
                        "description": "Brak uprawnień do aktualizacji tego zgłoszenia."
                    },
                    "404": {
                        "description": "Nie znaleziono zgłoszenia."
                    },
                    "422": {
                        "description": "Błąd walidacji."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Zgłoszenia"
                ],
                "summary": "Usuwa określone zgłoszenie",
                "description": "Trwale usuwa zgłoszenie z systemu. Dostępne w zależności od zdefiniowanej polityki autoryzacji (prawdopodobnie dla admina).",
                "operationId": "deleteTicket",
                "parameters": [
                    {
                        "name": "ticket",
                        "in": "path",
                        "description": "ID zgłoszenia",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Zgłoszenie pomyślnie usunięte."
                    },
                    "401": {
                        "description": "Błąd autoryzacji."
                    },
                    "403": {
                        "description": "Brak uprawnień do usunięcia tego zgłoszenia."
                    },
                    "404": {
                        "description": "Nie znaleziono zgłoszenia."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "title": "User",
                "description": "Model użytkownika",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Jan Kowalski"
                    },
                    "login": {
                        "type": "string",
                        "example": "jkowalski"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "j.kowalski@example.com"
                    },
                    "role": {
                        "type": "string",
                        "example": "user",
                        "enum": [
                            "user",
                            "it",
                            "admin"
                        ]
                    },
                    "active": {
                        "description": "Status aktywności konta.",
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
            "Ticket": {
                "title": "Ticket",
                "description": "Model zgłoszenia",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Błąd logowania"
                    },
                    "description": {
                        "type": "string",
                        "example": "Nie mogę się zalogować do systemu X."
                    },
                    "status": {
                        "type": "string",
                        "example": "nowe",
                        "enum": [
                            "nowe",
                            "w trakcie",
                            "zamknięte"
                        ]
                    },
                    "priority": {
                        "type": "string",
                        "example": "średni",
                        "enum": [
                            "niski",
                            "średni",
                            "wysoki"
                        ]
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "assigned_it_id": {
                        "type": "integer",
                        "example": 2,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
            "Message": {
                "title": "Message",
                "description": "Model wiadomości w zgłoszeniu",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "body": {
                        "type": "string",
                        "example": "Proszę o więcej szczegółów."
                    },
                    "ticket_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "sender_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
            "TicketFull": {
                "title": "TicketFull",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Ticket"
                    },
                    {
                        "properties": {
                            "user": {
                                "$ref": "#/components/schemas/User"
                            },
                            "assignedTo": {
                                "oneOf": [
                                    {
                                        "$ref": "#/components/schemas/User"
                                    }
                                ],
                                "nullable": true
                            },
                            "messages": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/MessageFull"
                                }
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "MessageFull": {
                "title": "MessageFull",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Message"
                    },
                    {
                        "properties": {
                            "sender": {
                                "$ref": "#/components/schemas/User"
                            }
                        },
                        "type": "object"
                    }
                ]
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}